% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_library.R
\name{clean_library}
\alias{clean_library}
\title{Clean up Mendeley bibtex library and save it somewhere else}
\usage{
clean_library(tex_file, bibtex_file, write_file, entry_remove = c("abstract",
  "file", "keywords", "url", "primaryClass", "eprint", "archivePrefix",
  "arxivId"))
}
\arguments{
\item{tex_file}{Path to tex_file which should be checked for citations}

\item{bibtex_file}{Path to Mendeley bibtex library}

\item{write_file}{Path and filename to which the new bibtex library is written}

\item{entry_remove}{Vector string that defines the entries in each reference to be removed}
}
\description{
Clean up Mendeley bibtex library and save it somewhere else
}
